Add WorkQueue - a general-purpose utility class for pushing work onto a queue for many worker threads to process (WIP - so far untested) This is like a generalised & enhanced version of ResourceBackgroundQueue; it supports: - concurrent tasks for any purpose, categorised by 'channels', with input / output data being passed around in an Any - registerable request / response handlers for any purpose, organised by 'channel' - multiple worker threads, which share the load for all channels - channels might be resource tasks, paging tasks, other domain-specific tasks - registering worker threads with the RenderSystem is optional (default is not to do this, so workers would not access RenderSystem resources)
The idea here is to have a small pool of shared background threads that can process many types of task, while still keeping the actual functionality neatly separated (in handlers). The more threaded processing we do, the more important it is to share threads for many purposes. ResourceBackgroundQueue could be reimplemented in terms of this new class later (note that when OGRE_THREAD_SUPPORT=1, the queue would need to register its threads with the render system, but not when OGRE_THREAD_SUPPORT=2.

--HG--
extra : convert_revision : svn%3A8631bf8a-c64b-0410-883c-f3eb003322f7/trunk%40821